==> 00. Importación de Librerías.ipynb <==
Para llevar el ancho de las celdas de Jupyter Notebooks al 100 % :

from IPython.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))  

### Generales:

import numpy as np
import pandas as pd
from scipy import stats
from scipy.integrate import simps
import math
import random
import string

### Para realizar gráficos:

import matplotlib.pyplot as plt
from ipywidgets import interact, fixed, IntSlider
import seaborn as sns
import bokeh.plotting as bk
from bokeh.palettes import brewer
import plotly.graph_objects as go
from matplotlib.pyplot import figure

### Para establecer los parámetros generales de un gráfico:

plt.style.use("classic")                # Establecemos el estilo (opción 1)
plt.style.use("default")                # Establecemos el estilo (opción 2) (Recomendado)
plt.rcParams['figure.figsize'] = (5,5)  # Establecemos el tamaño.

### Paquetes de Analisis de Datos y Machine Learning:

import lightgbm
import shap

### Paquetes de scikit-learn:

from sklearn.linear_model import (LinearRegression, Lasso, Ridge, ElasticNet)
from sklearn.model_selection import (KFold, ShuffleSplit, RepeatedKFold, TimeSeriesSplit)
from sklearn.model_selection import (train_test_split, cross_val_score, cross_validate, learning_curve, validation_curve) 
from sklearn.preprocessing import PolynomialFeatures
from sklearn import metrics
from sklearn.metrics import (confusion_matrix, roc_curve, precision_recall_curve, auc)
from sklearn import preprocessing
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import FunctionTransformer
from sklearn.compose import make_column_transformer
from sklearn.preprocessing import OneHotEncoder
from sklearn import linear_model
from sklearn import feature_extraction
from sklearn.neighbors import KNeighborsClassifier, KNeighborsRegressor
from sklearn.model_selection import GridSearchCV, RandomizedSearchCV
from sklearn.feature_selection import SelectKBest, SelectPercentile, RFE, f_regression, f_classif, chi2

### Paquetes de Finanzas y Análisis Técnico:

import ta as ta
from ta import add_all_ta_features

### Otros:

import datetime
import sys
import pickle
import time
from time import time
import os
from IPython.display import display, HTML
from functools import partial
